feat: add modal and seperate with custom hook (#240)
